+++ processor/src/test/fixtures/input/com/example/post_mount_or_update/PostMountOrUpdateInterface.java
@@ -1,0 +1,6 @@
+package com.example.post_mount_or_update;
+
+interface PostMountOrUpdateInterface
+{
+  void postMountOrUpdate();
+}