+++ processor/src/test/fixtures/input/com/example/post_mount_or_update/ProtectedAccessFromBasePostMountOrUpdateModel.java
@@ -1,0 +1,10 @@
+package com.example.post_mount_or_update;
+
+import com.example.post_mount_or_update.other.BaseProtectedAccessPostMountOrUpdateModel;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class ProtectedAccessFromBasePostMountOrUpdateModel
+  extends BaseProtectedAccessPostMountOrUpdateModel
+{
+}