+++ processor/src/test/fixtures/input/com/example/post_mount_or_update/PublicAccessViaInterfacePostMountOrUpdateModel.java
@@ -1,0 +1,23 @@
+package com.example.post_mount_or_update;
+
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.PostMountOrUpdate;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class PublicAccessViaInterfacePostMountOrUpdateModel
+  extends Component
+  implements PostMountOrUpdateInterface
+{
+  @PostMountOrUpdate
+  public void postMountOrUpdate()
+  {
+  }
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}