+++ processor/src/test/fixtures/input/com/example/post_mount_or_update/Suppressed2PublicAccessPostMountOrUpdateModel.java
@@ -1,0 +1,25 @@
+package com.example.post_mount_or_update;
+
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.PostMountOrUpdate;
+import react4j.annotations.ReactComponent;
+import react4j.annotations.SuppressReact4jWarnings;
+
+@ReactComponent
+abstract class Suppressed2PublicAccessPostMountOrUpdateModel
+  extends Component
+{
+  // This uses the CLASS retention suppression
+  @SuppressReact4jWarnings( "React4j:PublicLifecycleMethod" )
+  @PostMountOrUpdate
+  public void postMountOrUpdate()
+  {
+  }
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}